Safari-611.1.10/LayoutTests/fast/text/font-size-zero.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description">Tests that text with font-size:0 has zero height and width.</p>
<div id="console"></div>
<script>
var s = document.createElement('span');
s.style.fontSize = 0;
s.innerHTML = 'Text';
document.body.appendChild(s);

shouldBe("s.getBoundingClientRect().height", "0");
shouldBe("s.getBoundingClientRect().width", "0");

document.body.removeChild(s);
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
